requirements.txt
aiosqlite==0.19.0
anyio==4.0.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
ecdsa==0.18.0
fastapi==0.97.0
filelock==3.12.4
future==0.18.3
gevent==23.9.1
graphql-core==3.2.3
greenlet==3.0.0rc3
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-jose==3.3.0
requests==2.31.0
requests-file==1.5.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
strawberry-graphql==0.186.1
tldextract==3.6.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.22.0
zope.event==5.0
zope.interface==6.0
